// react-frontend/src/api/models/user-delivery-items-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OrderItemDTO } from './order-item-dto';

/**
 * Response containing user details and pending delivery items
 * @export
 * @interface UserDeliveryItemsResponse
 */
export interface UserDeliveryItemsResponse {
    /**
     * User\'s credential number/barcode
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'credentialNumber'?: string;
    /**
     * User\'s first name
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'firstName'?: string;
    /**
     * User\'s last name
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'lastName'?: string;
    /**
     * User\'s email
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'email'?: string;
    /**
     * User\'s residence
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'residence'?: string;
    /**
     * User\'s campus
     * @type {string}
     * @memberof UserDeliveryItemsResponse
     */
    'campus'?: string;
    /**
     * List of order items pending delivery for this user
     * @type {Array<OrderItemDTO>}
     * @memberof UserDeliveryItemsResponse
     */
    'pendingItems'?: Array<OrderItemDTO>;
    /**
     * Total number of items pending delivery
     * @type {number}
     * @memberof UserDeliveryItemsResponse
     */
    'totalItems'?: number;
}